+++ src/modules/Elsa.Workflows.Api/Endpoints/WorkflowInstances/Journal/ExecutionState/Models.cs
@@ -1,8 +1,8 @@
 using FastEndpoints;
 
-namespace Elsa.Workflows.Api.Endpoints.WorkflowInstances.Journal.HasUpdates;
+namespace Elsa.Workflows.Api.Endpoints.WorkflowInstances.Journal.ExecutionState;
 
-/// The request to check if there are updates for a workflow instance journal.
+/// The request to check the execution state of the workflow instance. 
 public class Request
 {
     /// The unique identifier of a workflow instance.
@@ -10,4 +10,4 @@ public class Request
 }
 
 /// Represents the response containing the last updated timestamp of a workflow instance.
-public record Response(DateTimeOffset UpdatedAt);
+public record Response(WorkflowStatus Status, WorkflowSubStatus SubStatus, DateTimeOffset UpdatedAt);
